FIX acc issue, use float to compute grouped_topk

+++ vllm/model_executor/layers/fused_moe/fused_moe.py
@@ -496,7 +496,7 @@ def grouped_topk(hidden_states: torch.Tensor,
     if scoring_func == "softmax":
         scores = torch.softmax(gating_output, dim=-1)
     elif scoring_func == "sigmoid":
-        scores = gating_output.sigmoid()
+        scores = gating_output.float().sigmoid()
     else:
         raise ValueError(f"Unsupported scoring function: {scoring_func}")
